tests/wpt/web-platform-tests/css/css-backgrounds/animations/border-image-slice-interpolation-stability.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>border-image-slice interpolation stability</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds-3/#border-image-slice">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="target"></div>
<script>
test(function() {
  var player = target.animate([
    {borderImageSlice: '50'},
    {borderImageSlice: '50'},
  ], {
    duration: 1,
    fill: 'forwards',
    easing: 'cubic-bezier(0, 1.5, 1, 1.5)',
  });
  player.pause();
  player.currentTime = 0.6345195996109396
  assert_equals(getComputedStyle(target).borderImageSlice, '50');
});
</script>
